requirements.txt
fastapi
flask
uvicorn
motor
google-generativeai
python-dotenv
pydantic
gunicorn
